#1166: modifications to get the GL reconciliation feature working, including porting a few of the low level ofbiz services from ofbiz R4 (opentaps 1.0) to financials, as the new version has added business logic and are no longer low level
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13605 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/financials/src/com/opensourcestrategies/financials/ledger/LedgerServices.java
@@ -3189,7 +3189,7 @@ public final class LedgerServices {
                 }
 
                 // need amount of trans entry for gl reconcile entry
-                Double amount = acctgTransEntry.getDouble("amount");
+                BigDecimal amount = acctgTransEntry.getBigDecimal("amount");
 
                 // prepare input for and call createGlReconciliationEntry
                 params = UtilMisc.toMap("glReconciliationId", glReconciliationId);
